+++ packages/xchain-cosmos/src/cosmos/types.ts
@@ -1,5 +1,4 @@
 import { TxParams } from '@xchainjs/xchain-client'
-import { BaseAmount } from '@xchainjs/xchain-util'
 import { BigSource } from 'big.js'
 import { Msg, PrivKey, codec } from 'cosmos-client'
 import { StdTxFee } from 'cosmos-client/api'
@@ -38,7 +37,6 @@ export type TransferOfflineParams = TransferParams & {
 }
 
 export type TxOfflineParams = TxParams & {
-  from_balance: BaseAmount
   from_account_number: string
   from_sequence: string
 }